Update setup script for docker and kubernetes

+++ scripts/helpers.sh
@@ -6,12 +6,18 @@ command_exists() {
 
 check_apt_packages() {
     if [ ! $? = 0 ] || [ ! "$(dpkg-query -W --show-format='${db:Status-Status}' "$@" 2>&1)" = installed ]; then
-        if command -v apt >/dev/null 2>&1; then
-            sudo apt install "$@" -y
-        else
-            sudo apt-get install "$@" -y
-        fi
+        if command_exists apt; then sudo apt install "$@" -y; else sudo apt-get install "$@" -y; fi
     fi
+}
+
+cleanup() {
+    if [ "$(dpkg-query -W --show-format='${db:Status-Status}' "$@" 2>&1)" = installed ]; then
+        if command_exists apt; then sudo apt purge "$@" 2>/dev/null; else sudo apt-get purge "$@" 2>/dev/null; fi
+    fi
+}
+
+update() {
+    if command_exists apt; then sudo apt update; else sudo apt-get update; fi
 }
 
 check_brew_packages() {
@@ -36,19 +42,19 @@ underline() {
 }
 
 error() {
-    printf '%s❌ Error: %s%s\n' "${FMT_BOLD}${FMT_RED}" "$*" "$FMT_RESET" >&2
+    printf '%s❌  Error: %s%s\n' "${FMT_BOLD}${FMT_RED}" "$*" "$FMT_RESET" >&2
 }
 
 success() {
-    printf '%s🎉 Success: %s%s\n' "${FMT_BOLD}${FMT_GREEN}" "$*" "$FMT_RESET" >&2
+    printf '%s🎉  Success: %s%s\n' "${FMT_BOLD}${FMT_GREEN}" "$*" "$FMT_RESET" >&2
 }
 
 warning() {
-    printf '%s👀 Warning: %s%s\n' "${FMT_YELLOW}" "$*" "$FMT_RESET" >&2
+    printf '%s👀  Warning: %s%s\n' "${FMT_YELLOW}" "$*" "$FMT_RESET" >&2
 }
 
 info() {
-    printf '%s👉 Info: %s%s\n' "${FMT_BLUE}" "$*" "$FMT_RESET" >&2
+    printf '%s👉  Info: %s%s\n' "${FMT_BLUE}" "$*" "$FMT_RESET" >&2
 }
 
 err_exit() {
@@ -58,7 +64,7 @@ err_exit() {
 
 checkyes() {
     local result=1
-    printf '%s🔆 %s [Y/n] %s' "${FMT_ORANGE}" "$*" "${FMT_RESET}" >&2
+    printf '%s🔆  %s [Y/n] %s' "${FMT_ORANGE}" "$*" "${FMT_RESET}" >&2
     read -r opt
     case $opt in
     y* | Y*) local result=0 ;;
